implement support for remote handset
