--- sandbox/test1.py
@@ -1,9 +1,0 @@
-import asyncio
-from bleak import BleakScanner
-
-async def main():
-    devices = await BleakScanner.discover()
-    for d in devices:
-        print(d.details)
-
-asyncio.run(main())
